nvme: LightNVM support
The first generation of Open-Channel SSDs will be based on NVMe. The
integration requires that a NVMe device register itself as a LightNVM
device.

The detection is through PCI IDs. The current supported are the
qemu nvme simulator and CNEX Labs Westlake SSD. The qemu nvme enables
support through vendor specific bits in the namespace identification and
the CNEX card is configured for LightNVM upon boot and does not need
detection other than PCI id.

After detection, vendor specific codes are used to identify the device
and enumerate supported features.

Signed-off-by: Javier González <jg@lightnvm.io>